requirements.txt
slack_bolt>=1.27.0
fastapi>=0.129.0
starlette>=0.52.1
uvicorn[standard]>=0.41.0
google-adk>=1.25.1
httpx>=0.28.1
python-dotenv>=1.2.1
aiohttp>=3.13.3
pytz>=2025.2
mcp>=1.0.0
